Make file_cache module into compound_cache into its own module
